requirements.txt
arxiv
sentence-transformers
numpy
pyyaml
requests
